requirements.txt
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
gunicorn==21.2.0
beautifulsoup4==4.13.0
requests==2.31.0
pymysql==1.1.0
cryptography==41.0.7

# Demand Forecasting Demo
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Inventory Optimization Demo
statsmodels>=0.14.0
lightgbm>=4.0.0
shap>=0.44.0

# Supplier Risk Assessment Demo
xgboost>=2.0.0
imbalanced-learn>=0.12.0
tensorflow>=2.15.0
